// Types/type.tsx
export enum AppScreen {
  Home = 'Home',
  SearchResult = 'SearchResult',
  SearchResultDetail = 'SearchResultDetail',
  FoodDetail = 'FoodDetail',
  MyList = 'MyList',
  Top10 = 'Top10',
}

export type StackParamsList = {
  Home: undefined
  SearchResult: undefined
  SearchResultDetail: undefined
  FoodDetail: undefined
  MyList: undefined
  Top10: undefined
}